% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_json.R
\name{aggregation_json}
\alias{aggregation_json}
\title{Aggregate data from Pendo}
\usage{
aggregation_json(auth_key)
}
\arguments{
\item{string}{input character vector}
}
\description{
Aggregate data from Pendo
}
