% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account.R
\name{get_account}
\alias{get_account}
\title{Retrieve specific account details}
\usage{
get_account(auth_key, account_id)
}
\arguments{
\item{string}{input character vector}
}
\description{
Retrieve specific account details
}
\examples{
get_account(some_auth_key, some_account_id)
}
