% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page.R
\name{get_pages_list}
\alias{get_pages_list}
\title{Retrieve list of pages from Pendo}
\usage{
get_pages_list(auth_key)
}
\arguments{
\item{string}{input character vector}
}
\description{
Retrieve list of pages from Pendo
}
\examples{
get_pages_list(some_auth_key)
}
