% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report.R
\name{get_report}
\alias{get_report}
\title{Retrieve specific report from Pendo in JSON format}
\usage{
get_report(auth_key, report_id)
}
\arguments{
\item{string}{input character vector}
}
\description{
Retrieve specific report from Pendo in JSON format
}
\examples{
get_report(some_auth_key, some_report_id)
}
