% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_visitor.R
\name{get_visitor}
\alias{get_visitor}
\title{Retrieve specific visitor from Pendo}
\usage{
get_visitor(auth_key, visitor_id)
}
\arguments{
\item{string}{input character vector}
}
\description{
Retrieve specific visitor from Pendo
}
\examples{
get_visitor(some_auth_key, some_report_id)
}
