% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_json.R
\name{report_json}
\alias{report_json}
\title{Retrieve specific report from Pendo on JSON format}
\usage{
report_json(auth_key, report_id)
}
\arguments{
\item{string}{input character vector}
}
\description{
Retrieve specific report from Pendo on JSON format
}
\examples{
report_json(some_auth_key, some_report_id)
}
